test source load

+++ tools/xrt/include/targets/file/FileTarget.h
@@ -6,6 +6,7 @@
 //-------------------------------------------------------------------------------------
 #pragma once
 
+#include <sys/types.h>
 #include <targets/common/Target.h>
 #include <fstream>
 
@@ -15,7 +16,7 @@ class FileTarget : public Target {
     std::ofstream out;
 
     inline void writeInstruction(uint32_t _instruction);
-    inline void writeInstruction(uint16_t _instruction, uint16_t _argument);
+    inline void writeInstruction(uint8_t _instruction, uint32_t _argument);
 public:
     FileTarget(std::string _path);
